+++ packages/sveltekit/src/index.ts
@@ -6,4 +6,3 @@ export { handleHooks } from "./server/hooks.js";
 export type GetSession = RequestEvent["locals"]["getSession"];
 export type GetSessionUser = RequestEvent["locals"]["getSessionUser"];
 export type SetSession = RequestEvent["locals"]["setSession"];
-export type ClearSession = RequestEvent["locals"]["clearSession"];