+++ test-apps/oauth/src/pages/api/google.ts
@@ -1,9 +1,10 @@
-import type { APIRoute } from "astro";
+import { AuthRequest } from "@lucia-auth/astro";
 import { auth, googleAuth } from "../../lib/lucia";
+import type { APIRoute } from "astro";
 
-export const get: APIRoute = async (request) => {
+export const get: APIRoute = async ({ url, request, cookies }) => {
 	const { existingUser, createUser, providerUser } = await googleAuth.validateCallback(
-		request.url.searchParams.get("code") || ""
+		url.searchParams.get("code") || ""
 	);
 	const user =
 		existingUser ||
@@ -11,12 +12,12 @@ export const get: APIRoute = async (request) => {
 			username: providerUser.name
 		}));
 	const session = await auth.createSession(user.userId);
-	const serializedCookies = auth.createSessionCookies(session);
+	const authRequest = new AuthRequest(auth, { request, cookies });
+	authRequest.setSession(session);
 	return new Response(null, {
 		status: 302,
 		headers: {
-			location: "/",
-			"set-cookie": serializedCookies.toString()
+			location: "/"
 		}
 	});
 };