+++ packages/express/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.1.2
+
+- [Breaking] Removed `locals.clearSession()`
+- `locals.setSession()` can be called using `null` to delete all session cookies
+- Update `lucia-auth` to v0.1.4
+
 ## 0.1.1
 
 - Add `locals.getSessionUser()` method